+++ src/core/os_wasm.c
@@ -42,7 +42,7 @@ static EM_BOOL onFocusChanged(int type, const EmscriptenFocusEvent* data, void* 
 
 static EM_BOOL onResize(int type, const EmscriptenUiEvent* data, void* userdata) {
   int newWidth, newHeight;
-  emscripten_webgl_get_drawing_buffer_size(state.context, &newWidth, &newHeight);
+  emscripten_get_canvas_element_size(CANVAS, &newWidth, &newHeight);
 
   if (state.width != (uint32_t) newWidth || state.height != (uint32_t) newHeight) {
     state.width = newWidth;
@@ -254,6 +254,23 @@ void os_set_clipboard_text(const char* text) {
   //
 }
 
+void* os_vm_init(size_t size) {
+  return malloc(size);
+}
+
+bool os_vm_free(void* p, size_t size) {
+  free(p);
+  return true;
+}
+
+bool os_vm_commit(void* p, size_t size) {
+  return true;
+}
+
+bool os_vm_release(void* p, size_t size) {
+  return true;
+}
+
 void os_thread_attach(void) {
   //
 }
@@ -306,7 +323,7 @@ bool os_window_open(const os_window_config* flags) {
 }
 
 bool os_window_is_open(void) {
-  return state.context > 0;
+  return true;
 }
 
 void os_window_get_size(uint32_t* width, uint32_t* height) {
@@ -315,10 +332,7 @@ void os_window_get_size(uint32_t* width, uint32_t* height) {
 }
 
 float os_window_get_pixel_density(void) {
-  int w, h, fw, fh;
-  emscripten_get_canvas_element_size(CANVAS, &w, &h);
-  emscripten_webgl_get_drawing_buffer_size(state.context, &fw, &fh);
-  return (w == 0 || h == 0) ? 0.f : (float) fw / w;
+  return 1.f; // TODO
 }
 
 void os_on_quit(fn_quit* callback) {